refactor: separated hotjar script to a function in utils
